fix(intellij): Multi instance IDE support
- Fixes problem while trying to open more than 1 angular based project.

- Fixes `open new project` into the same window

- Fixes race condition while opening toolWindow and server has been started yet

Now it uses one `Angular Console server` for all the IDE instances. Node Server
is started on the Application level with the first openned angular project and stopped with the
last one.

We added listener for the `jxBrowser` to  `onFinishLoadingFrame` which
give us `chrome-error:` message in case server is not started yet (this was the race condition problem)
which give possibility to reload the URL again until server is up.

Another major change is that we dont register ToolWindow statically inside
`plugin.xml`, but programatically inside the `NgWorkspaceMonitor` only after
server has been started and we can access URL.