feat(nx-mcp): additional tools for Nx Cloud integration
Sketches out tools for:
* Get CIPEs (search)
* Get CIPE details (single entity)
* Get runs (search)
* Get run details (single entity)
* Get tasks (search)
* Get task details (single entity)

Only CIPE tools are exposed at this stage so we can start dogfooding -- follow-ups are on the way for runs, tasks, and likely more.

Will also create a follow-up where we only expose the tools if the Cloud instance that we're trying to connect to can support it.